Integrated ratings data from pvpoke.com
The search list will now sort based off of the "overall" ranking of Pokemon in a given cup. Currently, still only gbl great, ultra, and master leagues are supported, but it will be easy to add more cups.

I intend to implement a bit more intelligence with the rankings to really guide the user into a good pvp team, starting with recommending "openers", then from there, likely good counters in the "overall" "switch" or "closer" rankings for that cup.