Team Analysis Type Coverage & Top Meta Threats
The team analysis screen will now display various type effectiveness visuals. Then below that is a list of the top 20 threats to the team that contain the highest rating in that cup's overall ratings list. In future development, implementing team battle simulations to find threats may be a good approach (PvPoke's design). For now, it appears that pulling from the already generated ratings list is actually quite helpful and may serve as a different tool that simulations all together.

Next in development is to create a widget for selecting team alternatives, which will naturally be counters to the current teams top threats. Additionally, I want to allow the user to tap a threat under the "Meta Threats" list, from which a list of counters will be available to choose. This may be nice when the user wants to counter a specific threat.