Battle Log Screen Implementation Started
The sort of "final" screen for the first production ready release has now been started. This "Battle Log" screen will allow the user to add any number of teams to a list, from which they can analyze for weaknesses. The purpose of this is to allow the user to assess the teams they engaged in a set of battles, to better understand the meta of their current standing in a given cup.