Team logger analysis
The user can now log any number of teams, edit them, and recieve a net type analysis of them. Additionally a list of the top 20 counters to the logged teams is presented, currently the abstractions between the logger and team builder do not allow for communication. It could be useful to edit a team in the team builder from the logger, given these usefully calculated counters. For now, the user can at least see these counters, and reflect the changes to a team manually in the team builder. It may actually be more intuitive to keep the to pages separated in logic, though it results in a subset of similar code.