Battle Log Implemented
The battle log allows the user to take a team from their team builder, and report wins, ties, and losses from opponent teams. Analysis can be done on on those opponent teams individually. Next I will build a way for the user to do a net analysis on the logged teams.