Logged Opponents Net Analysis
Perhaps the most useful and unique feature for this app. The user can now do a net analysis on all logged opponent teams, which will expose strengths, weaknesses, and top counters.

It would be nice to allow the user to select a counter Pokemon swap it into the working team.