Closer Integration to PvPoke's Ranking data
Pogo Teams now implicitly supports the GBL leagues, and their respective remixes. By providing a "openCups" field in the assets/gamemaster.json, Pogo Teams can read in the ranking information for any of the specified cups. This allows us to decide which cups should render in the team builder and rankings pages.

If the specified cup is not the standard 1500cp cap, the cup must be modified in the assets/gamemaster.json to with a cooresponding "cp" field.

Next in development is to persist the gamemaster and rankings data to firebase. This will allow us to make changes to the files, and all users will get updated information in (somewhat) realtime.