Test global api with GitHub Pages
I decided to update app model information to users by using a Pogo Teams GitHub Page as a sort of file bucket. I made a few test calls in main, and things are working. The idea is that there will be a time stamp on the GitHub Page, indicating whether the client needs to update their data (stored locally via hive - sqlite).

If they do need the update, the json information will be retrieved from the GitHub Page. This felt to me like the most reasonable free approach to providing updates to the app's model.