Fixed loading screen bug
I feel that the solution I found is not ideal, but it appears to be working fine. Upon startup, the app enters a material app with a loading screen. Once the loading is finished, that build returns the "true" material app where the app actually exists. Not sure if nesting material app widgets like this is good but I haven't encountered any significant bugs.