Removed Provider and fixed several startup bugs
The provider for the top layer of user teams was removed. It makes more sense to have this data read in from the db, and managed at the PogoScaffold level of the app.

The startup process is becoming more intuitive and bug free. The app should now be able to handle a variety of issues when attempting to update the Pokemon Rankings data.