Local cache implemented with Hive db.
This cache will hold all static Pogo data. And refreshes on app startup when the versionTimestamps collection has been modified in the cloud. A yet-to-be implemented feature in the app's settings will allow this cache to be refreshed by the user manually.

This will decrease the number of reads from Firebase for all static app data:
fast moves
charge moves
pokemon
cups
cup rankings